requirements.txt
flask==3.0.2
line-bot-sdk==3.9.0
pandas==2.2.1
python-dotenv==1.0.1
gunicorn==21.2.0 